More weapons to test

+++ src/aerial_attack.cpp
@@ -20,7 +20,7 @@ AerialAttack::AerialAttack(Stage& stage
 
 void AerialAttack::action(Worm* worm
 	, const b2Vec2& dest_pos
-	, const std::vector<float>& params) {
+	, const std::vector<int>& params) {
 
 	/*b2Body* b = worm->get_body();
