+++ src/baseball_bat.cpp
@@ -12,7 +12,7 @@ BaseballBat::BaseballBat(Stage& stage
 
 void BaseballBat::action(Worm* worm
 	, const b2Vec2& dest_pos
-	, const std::vector<float>& params) {
+	, const std::vector<int>& params) {
 
 	/*b2Body* b = worm->get_body();
