+++ src/teleportation.cpp
@@ -9,7 +9,7 @@ Teleportation::Teleportation(Stage& stage
 
 void Teleportation::action(Worm* worm
 	, const b2Vec2& dest_pos
-	, const std::vector<float>& params) {
+	, const std::vector<int>& params) {
 
 	/*b2Body* b = worm->get_body();
